Fix deleteById for PostService and remove ports exposed in docker-compose.yml

+++ user-api/src/main/java/hu/aut/bme/userapi/feature/users/service/impl/UserServiceImpl.java
@@ -47,6 +47,8 @@ class UserServiceImpl implements UserService {
 
 
     public void deleteUserById(int id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        }
     }
 }
